Reorganized folders, switched to Google Places for images, added extra functions for Teleport API
Changed the getImage() function to use Google's API. Added some extra
functions to the Teleport API to get city information and create
different files (csv, txt) for our locations. Added the cities, states,
and countries list which contains all locations that the Teleport API
references. Reorganized directory to show scripts that are currently not
being worked on in the API Test Scripts folder.